Fix to consider cell styles
Original solution skipped the styles, which caused date fields to be
returned as days since epoch time, instead of a formatted date.  There
was no way to determine which cells were actual numbers and which were
dates (w/o assuming).

Found a solution in
http://poi.apache.org/spreadsheet/how-to.html#xssf_sax_api, specifically
a link to
https://svn.apache.org/repos/asf/poi/trunk/src/examples/src/org/apache/poi/xssf/eventusermodel/XLSX2CSV.java.
Adapted that solution to your's to no longer ignore cell styles.

There is probably a more elegant solution, using the StreamingCell's
getCellType(), getStringCellValue(), getDateCellValue(), et. al.; but I
had to get this working in < 1 day.  So, I offer this as a temporary fix
until a more elegant solution can be found for using the other
StreamingCell methods/fields in the StreamingReader.