Fix NumberFormatException with malformed SST reference
Cells of type "s" (shared string reference) have an integer value that is a
lookup into the shared string table.

If a cell is of type "s" (shared string reference), and the value (<v> tag)
is blank, the StreamingExcelReader throws a NumberFormatException in
StreamingSheetReader.unformattedContents():

int idx = Integer.parseInt(lastContents);

This happens if you have a cell that looks like this:

<c r="L42" s="1" t="s"><v></v></c>

Normally, if Excel saves the file, that whole <c ...></c> wouldn't even be
there, but I've seen this happen twice now for data files from different
sources.

So unformattedContents() needs to check if the "lastContents" is null first.
This is effectively what the regular POI XSSF reader does - it checks
"isSetV()".