Add comments support to the accessible cells
Added option to allow reading the sheet comments and to allow
accessing them while processing the cells. The comments are read as lazy
as possible, but they cannot be read in a streaming fashion as they
are accessed by index. The default behaviour is not to read them. Added
an option to enable comments reading not to affect the current use of
the library.
Should be noted that if the input file uses the comments excessively they
may cause memory problems as with using XSSF reader.